--- src/main/java/com/example/blog7th/domain/UserRole.java
@@ -1,6 +1,0 @@
-package com.example.blog7th.domain; // 경로 끝에 세미콜론(;) 확인!
-
-public enum UserRole { // class를 enum으로 변경
-    USER,
-    ADMIN
-}
