+++ src/main/java/com/example/blog7th/domain/post/PostStatus.java
@@ -1,4 +1,4 @@
-package com.example.blog7th.domain;
+package com.example.blog7th.domain.post;
 
 public enum PostStatus {
     DRAFT,
