+++ src/main/java/com/example/blog7th/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.example.blog7th.repository;
 
-import com.example.blog7th.domain.Comment;
+import com.example.blog7th.domain.post.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
