+++ src/main/java/com/example/blog7th/repository/LikeRepository.java
@@ -1,6 +1,9 @@
 package com.example.blog7th.repository;
 
 import com.example.blog7th.domain.*;
+import com.example.blog7th.domain.post.Comment;
+import com.example.blog7th.domain.post.Post;
+import com.example.blog7th.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
