+++ src/main/java/com/example/blog7th/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.example.blog7th.repository;
 
-import com.example.blog7th.domain.Post;
+import com.example.blog7th.domain.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
